+++ src/app/products/guards/index.ts
@@ -1,1 +1,3 @@
 export * from './product-resolve.guard';
+export * from './product-exists.guard';
+export * from './products-state-preloading.guard';