+++ src/app/products/products-routing.module.ts
@@ -3,21 +3,19 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { ProductComponent, ProductListComponent, ProductViewComponent } from './components';
 import { ProductsComponent } from './products.component';
-import { ProductResolveGuard } from './guards/product-resolve.guard';
+import { ProductResolveGuard, ProductExistsGuard, ProductsStatePreloadingGuard } from './guards';
 
 const routes: Routes = [
   {
     path: 'product-list',
     component: ProductsComponent,
+    canActivate: [ProductsStatePreloadingGuard],
     children: [
       {
         path: 'product-view/:productID',
         component: ProductViewComponent,
-        resolve: { product: ProductResolveGuard }
-      },
-      {
-        path: 'product/:productID',
-        component: ProductComponent
+        resolve: { product: ProductResolveGuard },
+        canActivateChild: [ProductExistsGuard]
       },
       {
         path: '',
@@ -38,4 +36,4 @@ export class ProductsRoutingModule {
     ProductComponent,
     ProductViewComponent
   ];
- }
+}